+++ src/client/cypress/e2e/editor/waterIngress.cy.js
@@ -1,13 +1,28 @@
-import { createBorehole, createCompletion, startBoreholeEditing, loginAsAdmin } from "../helpers/testHelpers";
+import {
+  createBorehole,
+  createCompletion,
+  startBoreholeEditing,
+  loginAsAdmin,
+  createCasing,
+  handlePrompt,
+} from "../helpers/testHelpers";
 import { evaluateDisplayValue, setInput, setSelect } from "../helpers/formHelpers";
 import { addItem, startEditing, saveForm, deleteItem } from "../helpers/buttonHelpers";
 
 describe("Tests for the wateringress editor.", () => {
   it("Creates, updates and deletes wateringresses", () => {
-    // Precondition: Create casing to later link in observation
+    // Create borehole with completion and casing
     createBorehole({ "extended.original_name": "INTEADAL" })
       .as("borehole_id")
-      .then(id => createCompletion("testWateringress", id, 16000002, true))
+      .then(id =>
+        createCompletion("test wateringress", id, 16000002, true)
+          .as("completion_id")
+          .then(completionId => {
+            createCasing("casing-1", id, completionId, "2021-01-01", "2021-01-02", [
+              { fromDepth: 0, toDepth: 10, kindId: 25000103 },
+            ]);
+          }),
+      )
       .then(response => {
         expect(response).to.have.property("status", 200);
       });
@@ -22,25 +37,6 @@ describe("Tests for the wateringress editor.", () => {
 
     // start editing session
     startBoreholeEditing();
-
-    cy.get("[data-cy=completion-content-tab-casing]").click();
-    cy.wait("@casing_GET");
-
-    addItem("addCasing");
-    cy.wait("@codelist_GET");
-
-    setInput("name", "casing-1");
-    setInput("dateStart", "2021-01-01");
-    setInput("dateFinish", "2021-01-02");
-    setInput("casingElements.0.fromDepth", "0");
-    setInput("casingElements.0.toDepth", "10");
-    setSelect("casingElements.0.kindId", 2);
-    setSelect("casingElements.0.materialId", 3);
-    setInput("casingElements.0.innerDiameter", "3");
-    setInput("casingElements.0.outerDiameter", "4");
-
-    saveForm();
-    cy.wait("@casing_GET");
 
     cy.get('[data-cy="hydrogeology-menu-item"]').click({ force: true });
     cy.get('[data-cy="wateringress-menu-item"]').click({ force: true });
@@ -66,7 +62,7 @@ describe("Tests for the wateringress editor.", () => {
 
     evaluateDisplayValue("quantity", "viel (> 120 l/min)");
     evaluateDisplayValue("reliability", "fraglich");
-    evaluateDisplayValue("casingName", "testWateringress - casing-1");
+    evaluateDisplayValue("casingName", "test wateringress - casing-1");
 
     // edit wateringress
     startEditing();
@@ -75,10 +71,11 @@ describe("Tests for the wateringress editor.", () => {
     saveForm();
     evaluateDisplayValue("quantity", "mittel (30 - 120 l/min)");
     evaluateDisplayValue("conditions", "frei/ungespannt");
-    evaluateDisplayValue("casingName", "testWateringress - casing-1");
+    evaluateDisplayValue("casingName", "test wateringress - casing-1");
 
     // delete wateringress
     deleteItem();
+    handlePrompt("Wollen Sie diesen Eintrag wirklich löschen?", "Löschen");
     cy.wait("@wateringress_DELETE");
     cy.get("body").should("not.contain", "mittel (30 - 120 l/min)");
   });