+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 - Quality attribute for stratigraphy.
 - Support Amazon Cognito logout flow.
 - IsOpenBorehole option for instrumentation and backfill.
+- Support adding multiple field measurement results in `hydrogeology`.
 - Show prompt for unsaved changes when switching data card or completion tabs.
+- Add entry `resin` to codelist `backfill_material` and entries `U-probe` and `double U-probe` to codelist `casing_type`.
+- Show prompt before deleting data card.
 
 ### Changed